feat: draw sprites from top left corner
This allows to snap objects to the grid (even when the sprites
have odd size, since we don't need to divide by two to draw the
bounding box) and streamlines the "corner" management, assuming
the top left corner of the box will always be the origin.

En passant, this introduces some functionality to the test utils
to only render failed tests, to help with debugging.