feat: introduce text size
With this change, users can set a size for the string they are
rendering. There are only 4 sizes. We don't allow float scaling
for the simple reason we want to keep everything based on the
existing grid (i.e., chars will always be rendered with full
pixels, intead of fractions).