feat: convert tune to string and don't duplicate upon Continue
With this change we are changing the meaning of the Continue symbol:
previously it was meant as a placeholder which would make the parser
duplicate the previous Note.

Now, it's a symbol by itself and we interpret its meaning only
in the sequencer: whenever we see a Continue note we keep playing.