feat: kinematics with minimum displacement collision response
With this commit we get rid of dynamics in Latebit. The engine will
only support kinematics, will differentiate collisions and overlaps,
and the collision response will yield a displacement along the axis
that yields the minimum movement.

En passant, this change makes Camera responsible for following
objects instead of WorldManager.